backend/requirements.txt
Django
djangorestframework
djoser
Pillow
django-filter
gunicorn
psycopg2-binary
flake8
python-dotenv
